Created new class 'DataSet'
DataSet holds data imported from a single file. The plan is to have an instance of DataSet for each file opened, keep the raw data as private members, and then have methods on each class to evaluate data (according to thresholds etc) and spit out the information which will be plotted on the graph (bad counts, excitations). The information can then be concatenated (from e.g. several windows, each with their own instance of the class) and plotted.