Added method to handle changes in selected index of combo boxes
Need to work out how to disable one item in a combo box. Maybe remove it (or somehow keep it in the list but stop it from being shown). Or add a symbol at the beginning of the entry and check for that symbol then stop the user from choosing it.